+++ src/acat/azext_acat/aaz/latest/acat/report/_update.py
@@ -16,6 +16,12 @@ from azure.cli.core.aaz import *
 )
 class Update(AAZCommand):
     """Update a new AppComplianceAutomation report or update an exiting AppComplianceAutomation report.
+
+    :example: update acat report
+        az acat report update --reportName testName     --offer-guid "00000000-0000-0000-0000-000000000001"
+
+    :example: update resource list
+        az acat report update  --report-name youReportName  --resources test.json
     """
 
     _aaz_info = {