+++ src/aks-preview/azext_aks_preview/tests/latest/test_agentpool_decorator.py
@@ -29,7 +29,6 @@ from azure.cli.command_modules.acs._consts import (
     CONST_DEFAULT_NODE_VM_SIZE,
 )
 from azext_aks_preview._consts import (
-    CONST_DEFAULT_AUTOMATIC_SKU_NODE_VM_SIZE,
     CONST_DEFAULT_WINDOWS_NODE_VM_SIZE,
 )
 from azure.cli.command_modules.acs.agentpool_decorator import AKSAgentPoolParamDict
@@ -388,6 +387,60 @@ class AKSPreviewAgentPoolContextCommonTestCase(unittest.TestCase):
         ctx_2.attach_agentpool(agentpool_2)
         self.assertEqual(ctx_2.get_skip_gpu_driver_install(), None)
 
+    def common_get_driver_type(self):
+        # default
+        ctx_1 = AKSPreviewAgentPoolContext(
+            self.cmd,
+            AKSAgentPoolParamDict({"driver_type": None}),
+            self.models,
+            DecoratorMode.CREATE,
+            self.agentpool_decorator_mode,
+        )
+        self.assertEqual(ctx_1.get_driver_type(), None)
+        agentpool_1 = self.create_initialized_agentpool_instance(
+            gpu_profile=self.models.AgentPoolGPUProfile(
+                driver_type="CUDA"
+            )
+        )
+        
+        ctx_1.attach_agentpool(agentpool_1)
+        self.assertEqual(ctx_1.get_driver_type(), "CUDA")
+
+        # default
+        ctx_2 = AKSPreviewAgentPoolContext(
+            self.cmd,
+            AKSAgentPoolParamDict({"driver_type": None}),
+            self.models,
+            DecoratorMode.CREATE,
+            self.agentpool_decorator_mode,
+        )
+        self.assertEqual(ctx_2.get_driver_type(), None)
+        agentpool_2 = self.create_initialized_agentpool_instance(
+            gpu_profile=self.models.AgentPoolGPUProfile(
+                driver_type="GRID"
+            )
+        )
+        ctx_2.attach_agentpool(agentpool_2)
+        self.assertEqual(ctx_2.get_driver_type(), "GRID")
+
+        # custom
+        ctx_0 = AKSPreviewAgentPoolContext(
+            self.cmd,
+            AKSAgentPoolParamDict({"driver_type": "CUDA"}),
+            self.models,
+            DecoratorMode.CREATE,
+            self.agentpool_decorator_mode,
+        )
+        self.assertEqual(ctx_0.get_driver_type(), "CUDA")
+        agentpool_0 = self.create_initialized_agentpool_instance(
+            gpu_profile=self.models.AgentPoolGPUProfile(
+                driver_type=None
+            )
+        )
+
+        ctx_0.attach_agentpool(agentpool_0)
+        self.assertEqual(ctx_0.get_driver_type(), "CUDA")
+
     def common_get_os_sku(self):
         # default
         ctx_1 = AKSPreviewAgentPoolContext(
@@ -716,7 +769,7 @@ class AKSPreviewAgentPoolContextCommonTestCase(unittest.TestCase):
             DecoratorMode.CREATE,
             self.agentpool_decorator_mode,
         )
-        self.assertEqual(ctx_5.get_node_vm_size(), CONST_DEFAULT_AUTOMATIC_SKU_NODE_VM_SIZE)
+        self.assertEqual(ctx_5.get_node_vm_size(), "")
 
     def common_get_gateway_prefix_size(self):
         # default
@@ -835,6 +888,9 @@ class AKSPreviewAgentPoolContextStandaloneModeTestCase(
 
     def test_get_skip_gpu_driver_install(self):
         self.common_get_skip_gpu_driver_install()
+
+    def test_get_driver_type(self):
+        self.common_get_driver_type()
 
     def test_get_enable_secure_boot(self):
         self.common_get_enable_secure_boot()