+++ src/monitor-pipeline-group/azext_monitor_pipeline_group/aaz/latest/monitor/pipeline_group/_delete.py
@@ -20,9 +20,9 @@ class Delete(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2023-10-01-preview",
+        "version": "2024-10-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.monitor/pipelinegroups/{}", "2023-10-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.monitor/pipelinegroups/{}", "2024-10-01-preview"],
         ]
     }
 
@@ -49,7 +49,7 @@ class Delete(AAZCommand):
             required=True,
             id_part="name",
             fmt=AAZStrArgFormat(
-                pattern="^(?!-)[a-zA-Z0-9-]{3,10}[^-]$",
+                pattern="^(?!-)[a-zA-Z0-9-]{3,32}[^-]$",
             ),
         )
         _args_schema.resource_group = AAZResourceGroupNameArg(
@@ -143,7 +143,7 @@ class Delete(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-10-01-preview",
+                    "api-version", "2024-10-01-preview",
                     required=True,
                 ),
             }